Factored out the relay-server code in a separate class.
Why? This makes it easier to build tests which depend on both, the
relay-server and client.
