+++ src/go/cmd/http-relay-server/server/BUILD
@@ -1,18 +1,20 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_binary", "go_library")
-load("@rules_oci//oci:defs.bzl", "oci_image")
-load("@rules_pkg//pkg:tar.bzl", "pkg_tar")
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 package(default_visibility = ["//visibility:public"])
 
 go_library(
     name = "go_default_library",
     srcs = [
-        "main.go",
+        "backend_context.go",
+        "broker.go",
+        "handler_single.go",
+        "handler_stream.go",
+        "http_proto.go",
+        "server.go",
+        "utils.go",
     ],
-    importpath = "github.com/googlecloudrobotics/core/src/go/cmd/http-relay-server",
-    visibility = ["//visibility:private"],
+    importpath = "github.com/googlecloudrobotics/core/src/go/cmd/http-relay-server/server",
     deps = [
-        "//src/go/cmd/http-relay-server/server:go_default_library",
         "//src/proto/http-relay:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@com_github_prometheus_client_golang//prometheus:go_default_library",
@@ -28,19 +30,18 @@ go_library(
     ],
 )
 
-go_binary(
-    name = "http-relay-server-app",
+go_test(
+    name = "go_default_test",
+    size = "small",
+    srcs = [
+        "broker_test.go",
+        "server_test.go",
+    ],
     embed = [":go_default_library"],
-)
-
-pkg_tar(
-    name = "http-relay-server-image-layer",
-    srcs = [":http-relay-server-app"],
-)
-
-oci_image(
-    name = "http-relay-server-image",
-    base = "@distroless_base",
-    entrypoint = ["/http-relay-server-app"],
-    tars = [":http-relay-server-image-layer"],
+    visibility = ["//visibility:private"],
+    deps = [
+        "//src/proto/http-relay:go_default_library",
+        "@com_github_getlantern_httptest//:go_default_library",
+        "@org_golang_google_protobuf//proto:go_default_library",
+    ],
 )
