+++ src/go/cmd/http-relay-server/server/backend_context.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"crypto/rand"
