+++ src/go/cmd/http-relay-server/server/handler_single.go
@@ -1,4 +1,4 @@
-package main
+package server
 
 import (
 	"context"
