+++ src/go/cmd/http-relay-server/server/http_proto.go
@@ -1,5 +1,5 @@
 // Utilities for proto conversion
-package main
+package server
 
 import (
 	"log"
